♻️ Refactor `UriToStringBehavior` to a service
This commit will refactor the `UriToStringBehavior` to make it a service
object instead.
