+++ app/indexers/collection_indexer.rb
@@ -4,15 +4,16 @@ class CollectionIndexer < Hyrax::CollectionIndexer
   # This indexes the default metadata. You can remove it if you want to
   # provide your own metadata and indexing.
   include Hyrax::IndexesBasicMetadata
-  include UriToStringBehavior
 
   # Uncomment this block if you want to add custom indexing behavior:
   def generate_solr_document
     super.tap do |solr_doc|
-      solr_doc["creator_sim"] = solr_doc["creator_tesim"] = convert_uri_to_value(['creator'])
-      solr_doc["subject_sim"] = solr_doc["subject_tesim"] = convert_uri_to_value(['subject'])
-      solr_doc["contributor_sim"] = solr_doc["contributor_tesim"] = convert_uri_to_value(['contributor'])
-      solr_doc["language_sim"] = solr_doc["language_tesim"] = convert_uri_to_value(['language'])
+      converter = UriToStringConverterService.new(object)
+
+      solr_doc["creator_sim"] = solr_doc["creator_tesim"] = converter.convert_uri_to_value(['creator'])
+      solr_doc["subject_sim"] = solr_doc["subject_tesim"] = converter.convert_uri_to_value(['subject'])
+      solr_doc["contributor_sim"] = solr_doc["contributor_tesim"] = converter.convert_uri_to_value(['contributor'])
+      solr_doc["language_sim"] = solr_doc["language_tesim"] = converter.convert_uri_to_value(['language'])
       solr_doc["bulkrax_identifier_sim"] = object.bulkrax_identifier
       solr_doc["account_cname_tesim"] = Site.instance&.account&.cname
       solr_doc[CatalogController.title_field] = Array(object.title).first
